requirements.txt
flask
fastapi
tensorflow-cpu
pillow
numpy<2